+++ app/src/main/java/com/smashing/app/presentation/write/model/MatchPlayer.kt
@@ -1,4 +1,4 @@
-package com.smashing.app.presentation.submit.model
+package com.smashing.app.presentation.write.model
 
 data class MatchPlayer(
     val userId: String,
