+++ app/src/main/res/drawable/ic_check_button.xml
@@ -1,0 +1,14 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="40dp"
+    android:height="40dp"
+    android:viewportWidth="40"
+    android:viewportHeight="40">
+  <path
+      android:strokeWidth="1"
+      android:pathData="M20,0.5C22.699,0.5 25.231,1.013 27.602,2.035C29.982,3.062 32.045,4.451 33.796,6.203C35.547,7.955 36.937,10.019 37.967,12.398C38.992,14.767 39.504,17.299 39.5,19.999C39.496,22.7 38.984,25.233 37.965,27.603C36.942,29.981 35.553,32.044 33.797,33.796C32.041,35.548 29.977,36.939 27.602,37.967C25.237,38.99 22.704,39.503 20,39.5C17.299,39.5 14.766,38.987 12.398,37.965C10.021,36.938 7.957,35.548 6.203,33.796C4.45,32.044 3.06,29.981 2.035,27.603C1.014,25.233 0.501,22.7 0.5,20C0.499,17.3 1.011,14.768 2.035,12.398C3.063,10.019 4.453,7.955 6.204,6.203C7.955,4.451 10.018,3.062 12.398,2.035C14.769,1.013 17.301,0.5 20,0.5Z"
+      android:fillColor="#2261FF"
+      android:strokeColor="#2261FF"/>
+  <path
+      android:pathData="M18.446,22.036L25.571,15.241C25.739,15.08 25.935,15 26.159,15C26.383,15 26.58,15.08 26.748,15.241C26.916,15.401 27,15.591 27,15.812C27,16.033 26.916,16.223 26.748,16.383L19.034,23.76C18.866,23.92 18.67,24 18.446,24C18.222,24 18.026,23.92 17.858,23.76L14.243,20.312C14.074,20.151 13.994,19.961 14,19.741C14.007,19.521 14.095,19.33 14.264,19.169C14.432,19.008 14.632,18.928 14.863,18.929C15.094,18.929 15.293,19.009 15.462,19.169L18.446,22.036Z"
+      android:fillColor="#ffffff"/>
+</vector>
