+++ app/src/main/res/drawable/ic_check_button_default.xml
@@ -1,0 +1,11 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="30dp"
+    android:height="30dp"
+    android:viewportWidth="30"
+    android:viewportHeight="30">
+  <path
+      android:strokeWidth="1"
+      android:pathData="M15,0.5C17.008,0.5 18.89,0.882 20.652,1.642C22.424,2.406 23.958,3.439 25.259,4.741C26.56,6.044 27.594,7.578 28.36,9.349C29.122,11.109 29.503,12.991 29.5,14.999C29.497,17.008 29.115,18.891 28.358,20.652C27.597,22.422 26.565,23.957 25.26,25.259C23.954,26.562 22.419,27.595 20.651,28.36C18.894,29.121 17.012,29.502 15,29.5C12.99,29.5 11.108,29.119 9.349,28.359C7.579,27.595 6.045,26.561 4.741,25.259C3.438,23.956 2.405,22.423 1.642,20.652C0.883,18.892 0.501,17.009 0.5,15C0.499,12.991 0.88,11.109 1.641,9.349C2.406,7.578 3.44,6.044 4.741,4.741C6.042,3.439 7.576,2.406 9.348,1.642C11.109,0.882 12.992,0.5 15,0.5Z"
+      android:fillColor="#576071"
+      android:strokeColor="#576071"/>
+</vector>
